backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.3
SQLAlchemy==2.0.30
alembic==1.13.2
httpx==0.27.0
qrcode[pil]==7.4.2
